refactor: Move `constants.ts` into `core`
